+++ src/Services/FaraImporter.php
@@ -26,6 +26,18 @@ class FaraImporter
         'stattrafficincome.csv' => 'fara_stat_traffic_income',
     ];
 
+    /**
+     * Columns to be ignored.
+     *
+     * @var array
+     */
+    protected $columnFilter = [
+        'fara_basic_journey' => ['remarkida', 'remark2ida', 'packagetour', 'announcementtype', 'tariffid', 'tarifftype'],
+        'fara_basic_line' => ['tariffid'],
+        'fara_basic_stop' => ['stopshortname', 'tariffzone', 'xcoordinate', 'ycoordinate'],
+        'fara_company' => ['companyshortname', 'csurl', 'address', 'postalcodeid', 'accountnumber'],
+    ];
+
     public function __construct()
     {
         $this->logPrintfInit('[FaraImporter]: ');
@@ -37,14 +49,17 @@ class FaraImporter
         if (($handle = fopen($file, 'r')) !== false) {
             $table = $this->localTables[basename($file)];
             $feeder = new DbBulkInsert($table, 'upsert');
-            $dbCols = fgetcsv($handle);
+            $dbCols = $this->getFilteredColumns($table, fgetcsv($handle));
             $feeder->unique($dbCols);
             while (($values = fgetcsv($handle)) !== false) {
                 foreach ($values as $k => $val) {
                     $value = trim($val, '"');
-                    if (strlen($value) === 0) $values[$k] = null;
+                    if (strlen($value) === 0) {
+                        $values[$k] = null;
+                    }
                 }
-                $feeder->addRecord(array_combine($dbCols, $values));
+                $dbVals = array_filter($values, fn($key) => array_key_exists($key, $dbCols), ARRAY_FILTER_USE_KEY);
+                $feeder->addRecord(array_combine($dbCols, $dbVals));
                 $records += 1;
             }
             fclose($handle);
@@ -52,6 +67,18 @@ class FaraImporter
         }
         $this->debug('%s: Imported %d records', basename($file), $records);
         return $records;
+    }
+
+    // Remove column names as specified by the columnFilter array without re-
+    // indexing the returned array.
+    protected function getFilteredColumns(string $tableName, array $columns): array
+    {
+        if (!array_key_exists($tableName, $this->columnFilter)) {
+            // No column filter available for this table.
+            return $columns;
+        }
+        $filter = $this->columnFilter[$tableName];
+        return array_filter($columns, fn($col) => !in_array($col, $filter));
     }
 
     public function deleteImport(string $id)